Add InstructionService and its corresponding tests
The InstructionService has been added, with methods for creating instructions based on a configuration. Each configuration is represented by a sorted dictionary with the color of each item as the value. InstructionServiceTests have also been created to test different scenarios and configurations. Furthermore, the values of enum Color were reordered, and Class1 was renamed to InstructionService. Classes named TheoryData, SerialServiceTests and TheoryData2 were also modified accordingly.