Add comments to Color enum
Detailed comments were added to the Color enum in the Sprinti.Domain namespace. The objective is to clarify its purpose and underscore its important role in indexing the color-depot in the default/reset state. The comments also serve as a warning against unnecessary changes unless the color storage formation undergoes modification.