Add file writing to VideoStream and set up Nix flake
In VideoStream.cs, file writing operations were added to the code using the new FileStorage instance. In addition, logging functionality was improved to provide information on the received image dimensions. Furthermore, the project now includes a basic flake setup with `flake.nix` and `cvsharp.nix` for package management. The ignored files list is also updated to exclude build artifacts. A reference to OpenCvSharp4 for Ubuntu was added to the Sprinti API project file.