Add capture interval option for video streaming
A new configuration option for capturing interval in seconds has been added to the StreamOptions. Additionally, there has been an adjustment in VideoStream.cs to accommodate this new feature. The capturing of images now occurs depending on the specified interval, and the timestamp has been moved to the image file name for easier tracking.