Update RunInstructions method to RunInstructionsAndFinish
The RunInstructions method in SerialService has been updated and renamed to RunInstructionsAndFinish. This method now includes the process of starting and finishing the sequence of commands within it. This change led to a simplification of the command sequence created in the InstructionService and the removal of the related tests in SerialService and InstructionService tests.