Add new features and update package dependencies
This commit encompasses addition of Iot.Device.Bindings and System.Device.Gpio packages as dependencies and removal of the unnecessary ones. It also includes the implementation of RunInstructionsAndFinish method to the SerialController and creates ButtonService for managing button operations. The change in the IButtonService interface location from WorkflowService file to its own exclusive file is also incorporated. Updates were mirrored in the Tests folder as well.