Refactor code for clean checkout
Significant code clean-up was performed to ensure a tidy and error-free checkout. Changes made include adjustment of whitespaces and newlines, and reordering of variables and references for better readability. Additionally, unnecessary code that was found to be redundant or not in use was also removed. Program URL was also updated.