Add ImageSelector class and update VideoStream
A new class, ImageSelector, has been introduced that allows the extraction of a mask from an image based on HSV color range. An instance of this is now used in VideoStream to isolate blue colors in the flow. Furthermore, the Nix package cvsharp has been updated to support GTK, resolving a previously known issue.