Update cube detection functionality
Cube detection has been updated to include an optional show parameter in the DetectCubes method, allowing the user to visualize the detection process. A private helper function 'Show' was added to the CubeDetector class to enable this visualization. Furthermore, the lookup table referenced in CubeDetectorTests has been modified to use the second config instead of the first one.