Refactor WorkflowService and reduce dependencies
The WorkflowService has been refactored to use a different service, the 'IVideoProcessor', instead of the 'IVideoService'. This resulted in a change in the process to run the detection and handle its result. Unnecessary dependencies in the 'deps.nix' file have also been removed, making the project leaner. Lastly, the stream feature in the app settings has been enabled.