Delete "config.json" and revise references in "Startup.cs"
The configuration file "config.json" has been deleted, requiring modifications to the references in "Startup.cs" code. The path was updated to point to a new file "detection.json". Additionally, the deserializing process was revised to accommodate the new data structure, focusing explicitly on the "LookupConfigs" within the "Detection" object of the new JSON file. A supplementary class, "DetectionJson", was created to represent the new JSON structure.