Implement detection duration measurement in StreamController
Added a Stopwatch in `StreamController.cs` to measure the time it takes for `RunDetection` function to execute. Returns execution time as part of a new `RunDetectionDto` object which also includes the original `CubeConfig`. Also involves some tests' readability enhancements and includes a new dataset to `detection.json` file testing purposes.