Add option to enable or disable modules
Implemented the ISprintiOptions interface to allow toggling of different modules from the configuration settings. Added the 'Enabled' property in 'ISprintiOptions' and used it in various modules to control whether they are added to the service collection or not. Updated all modules to check their respective 'Enabled' settings before adding services and added 'Enabled' property to all necessary options classes.