Add SkiaSharpAdapter and Fontconfig dependency
The SkiaSharpAdapter is added to ModuleRegistry.cs to enhance graphic capabilities. This involves registering SkiaSharpAdapter, which corresponds to a GitHub pull request pointed out in the comments. The Fontconfig module, a library designed to provide a list of available fonts to applications, has been added to the Sprinti's Nix package configuration, further extending the application's functionality.