Update comment links and tweak module format
The code was updated to provide more accurate comment references. A link pointing to a GitHub issue comment was introduced in the ModuleRegistry file. A new link that leads to a discourse.nixos.org discussion was added to the sprinti.nix file. Also, minor syntax adjustments were made to the module configuration for better readability.