Refactor WorkflowProcedure to use IInstructionService
In WorkflowService.cs, the order of asynchronous tasks has been adjusted for better efficiency. In SerialController.cs, changes were made to accommodate the IInstructionService which provides a more streamlined and flexible method for generating instructions. Instructions are now defined as a sequence, offering increased maintainability and scalability of the code.