Refactor workflow power unit and rename related classes
Changes were made to the Workflow and Serial Services to accommodate the power unit conversion from Watt-hours to NanoJoules. The unit conversion is now done directly in the method RunWorkflowProcedure for more accurate energy calculation. The WorkflowController was deleted; StreamController was renamed to DetectionController to better represent its functionality. The StreamInstructController and SerialServiceTests classes were updated accordingly.